+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 Changelog
 =========
 
-2.0.0-beta1 (2017-09-29)
+2.0.0-beta4 (2017-MM-DD)
+------------------------
+
+[Full Changelog](https://github.com/contao-bootstrap/layout/compare/2.0.0-beta3...2.0.0-beta4)
+
+Implemented enhancements:
+ 
+ - Rewrite `rounded` and `rounded-*` classes to the picture element.
+
+
+2.0.0-beta3 (2017-09-29)
 ------------------------
 
 [Full Changelog](https://github.com/contao-bootstrap/layout/compare/2.0.0-beta2...2.0.0-beta3)